refactor: Extract calculateMembershipExpirationDate function to strip_helper
