+++ utils/server/stripe/handleCheckoutSessionCompleted.ts
@@ -14,6 +14,7 @@ import {
   userProfileQuery,
 } from '../supabase';
 import {
+  calculateMembershipExpirationDate,
   updateUserAccountByEmail,
   updateUserAccountById,
 } from './strip_helper';
@@ -133,32 +134,6 @@ export default async function handleCheckoutSessionCompleted(
       });
     }
   })();
-}
-
-async function calculateMembershipExpirationDate(
-  planGivingWeeks: string | undefined,
-  planCode: string | undefined,
-  sessionCreatedDate: Date,
-): Promise<Date | undefined> {
-  const previousDate = dayjs(sessionCreatedDate || undefined);
-  // If has planGivingWeeks, use it to calculate the expiration date
-  if (planGivingWeeks && typeof planGivingWeeks === 'string') {
-    return previousDate.add(+planGivingWeeks, 'week').toDate();
-  }
-  // else extend the expiration date based on the plan code
-  else if (planCode === PaidPlan.ProOneTime) {
-    return previousDate.add(1, 'month').toDate();
-  } else if (planCode === PaidPlan.ProMonthly) {
-    return previousDate.add(1, 'month').toDate();
-  } else if (planCode === PaidPlan.UltraOneTime) {
-    return previousDate.add(1, 'month').toDate();
-  } else if (planCode === PaidPlan.UltraMonthly) {
-    return previousDate.add(1, 'month').toDate();
-  } else if (planCode === PaidPlan.UltraYearly) {
-    return previousDate.add(1, 'year').toDate();
-  }
-  // Return undefined if no conditions are met
-  return undefined;
 }
 
 async function addCreditToUser(
